// types/errors.ts
export enum Errors {
  MISSING_SEND_MESSAGES = "MISSING_SEND_MESSAGES",
  MISSING_MANAGE_ROLES = "MISSING_MANAGE_ROLES",
  MISSING_KICK_MEMBERS = "MISSING_KICK_MEMBERS",
  MISSING_VIEW_CHANNEL = "MISSING_VIEW_CHANNEL",
  MISSING_READ_MESSAGE_HISTORY = "MISSING_READ_MESSAGE_HISTORY",
  MISSING_MANAGE_NICKNAMES = "MISSING_MANAGE_NICKNAMES",
  MISSING_MUTE_MEMBERS = "MISSING_MUTE_MEMBERS",
  MISSING_DEAFEN_MEMBERS = "MISSING_DEAFEN_MEMBERS",
  MISSING_SEND_TTS_MESSAGE = "MISSING_SEND_TTS_MESSAGE",
  MISSING_MANAGE_MESSAGES = "MISSING_MANAGE_MESSAGES",
  MISSING_MANAGE_CHANNELS = "MISSING_MANAGE_CHANNELS",
  MISSING_CREATE_INSTANT_INVITE = "MISSING_CREATE_INSTANT_INVITE",
  MISSING_MANAGE_WEBHOOKS = "MISSING_MANAGE_WEBHOOKS",
  MISSING_MANAGE_EMOJIS = "MISSING_MANAGE_EMOJIS",
  MISSING_BAN_MEMBERS = "MISSING_BAN_MEMBERS",
  MISSING_MANAGE_GUILD = "MISSING_MANAGE_GUILD",
  MISSING_VIEW_AUDIT_LOG = "MISSING_VIEW_AUDIT_LOG",
  DELETE_MESSAGES_MIN = "DELETE_MESSAGES_MIN",
  DELETE_MESSAGES_MAX = "DELETE_MESSAGES_MAX",
  MESSAGE_MAX_LENGTH = "MESSAGE_MAX_LENGTH",
  NICKNAMES_MAX_LENGTH = "NICKNAMES_MAX_LENGTH",
  PRUNE_MIN_DAYS = "PRUNE_MIN_DAYS"
}